chore: create new file for each Constant struct
