+++ Pharus/Coordinator/Service/Constants/Assets.swift
@@ -1,0 +1,13 @@
+//
+//  Assets.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct Assets {
+     let icons = Icons()
+     let images = Images()
+}
