+++ Pharus/Coordinator/Service/Constants/Avatar.swift
@@ -1,0 +1,13 @@
+//
+//  Avatar.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct Avatar {
+    let fullImage = FullImage()
+    let circleImage = CircleImage()
+}
