+++ Pharus/Coordinator/Service/Constants/CircleImage.swift
@@ -1,0 +1,32 @@
+//
+//  CircleImage.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct CircleImage {
+     let avatar1 = "avatar1CircleImage"
+     let avatar2 = "avatar2CircleImage"
+     let avatar3 = "avatar3CircleImage"
+     let avatar4 = "avatar4CircleImage"
+     let avatar5 = "avatar5CircleImage"
+     let avatar6 = "avatar6CircleImage"
+     let avatar7 = "avatar7CircleImage"
+     let avatar8 = "avatar8CircleImage"
+     let avatar9 = "avatar9CircleImage"
+     let avatar10 = "avatar10CircleImage"
+     let avatar11 = "avatar11CircleImage"
+     let avatar12 = "avatar12CircleImage"
+     let avatar13 = "avatar13CircleImage"
+     let avatar14 = "avatar14CircleImage"
+     let avatar15 = "avatar15CircleImage"
+     let avatar16 = "avatar16CircleImage"
+     let avatar17 = "avatar17CircleImage"
+     let avatar18 = "avatar18CircleImage"
+     let avatar19 = "avatar19CircleImage"
+     let avatar20 = "avatar20CircleImage"
+     let suffix = "CircleImage"
+}
