+++ Pharus/Coordinator/Service/Constants/Company.swift
@@ -1,0 +1,16 @@
+//
+//  Company.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct Company {
+     let havaianasLogoImage = "havaianasLogoImage"
+     let ioasysLogoImage = "ioasysLogoImage"
+     let osklenLogoImage = "osklenLogoImage"
+     let dupeLogoImage = "dupéLogoImage"
+     let alpargatasLogoImage = "alpargatasLogoImage"
+}
