+++ Pharus/Coordinator/Service/Constants/FullImage.swift
@@ -1,0 +1,32 @@
+//
+//  FullImage.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct FullImage {
+     let avatar1 = "avatar1FullImage"
+     let avatar2 = "avatar2FullImage"
+     let avatar3 = "avatar3FullImage"
+     let avatar4 = "avatar4FullImage"
+     let avatar5 = "avatar5FullImage"
+     let avatar6 = "avatar6FullImage"
+     let avatar7 = "avatar7FullImage"
+     let avatar8 = "avatar8FullImage"
+     let avatar9 = "avatar9FullImage"
+     let avatar10 = "avatar10FullImage"
+     let avatar11 = "avatar11FullImage"
+     let avatar12 = "avatar12FullImage"
+     let avatar13 = "avatar13FullImage"
+     let avatar14 = "avatar14FullImage"
+     let avatar15 = "avatar15FullImage"
+     let avatar16 = "avatar16FullImage"
+     let avatar17 = "avatar17FullImage"
+     let avatar18 = "avatar18FullImage"
+     let avatar19 = "avatar19FullImage"
+     let avatar20 = "avatar20FullImage"
+     let suffix = "FullImage"
+}
