+++ Pharus/Coordinator/Service/Constants/Icons.swift
@@ -1,0 +1,41 @@
+//
+//  Icons.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct Icons {
+    let backArrowIcon = "backArrowIcon"
+    let logoutIcon = "logoutIcon"
+    let bookIcon = "bookIcon"
+    let clockIcon = "clockIcon"
+    let rulesIcon = "rulesIcon"
+    let feedbackIcon = "feedbackIcon"
+    let xmarkIcon = "xmarkIcon"
+    let checkboxCheckmarkButtonIcon = "checkboxCheckmarkIcon"
+    let expandIcon = "ExpandIcon"
+    let envelopeIcon = "envelopeIcon"
+    let notificationEnvelopeIcon = "notificationEnvelopeIcon"
+    let uploadIcon = "uploadIcon"
+    let checkIcon = "checkIcon"
+    let rightArrowIcon = "rightArrowIcon"
+    let warningIcon = "warningIcon"
+    let medalIcon =  "medalIcon"
+    let companyIcon = "companyIcon"
+    let errorIcon = "errorIcon"
+
+    let avatarTabBarIcon = "avatarTabBarIcon"
+    let homeTabBarIcon = "homeTabBarIcon"
+    let profileTabBarIcon = "profileTabBarIcon"
+    let projectsTabBarIcon = "projectsTabBarIcon"
+    let rankingTabBarIcon = "rankingTabBarIcon"
+
+    let avatarTabBarIconSelected = "avatarTabBarIconSelected"
+    let homeTabBarIconSelected = "homeTabBarIconSelected"
+    let profileTabBarIconSelected = "profileTabBarIconSelected"
+    let projectsTabBarIconSelected = "projectsTabBarIconSelected"
+    let rankingTabBarIconSelected = "rankingTabBarIconSelected"
+}
