+++ Pharus/Coordinator/Service/Constants/Images.swift
@@ -1,0 +1,27 @@
+//
+//  Images.swift
+//  Pharus
+//
+//  Created by Victor Colen on 15/06/22.
+//
+
+import Foundation
+
+struct Images {
+     let avatar = Avatar()
+     let company = Company()
+
+     let appLogoImage = "PharusImage"
+
+     let firstPlaceImage = "1stPlaceImage"
+     let secondPlaceImage = "2ndPlaceImage"
+     let thirdPlaceImage = "3rdPlaceImage"
+     let fourthPlaceImage = "4thPlaceImage"
+     let fifthPlaceImage = "5thPlaceImage"
+
+     let firstPlaceMedalImage = "1stPlaceMedalImage"
+     let secondPlaceMedalImage = "2ndPlaceMedalImage"
+     let thirdPlaceMedalImage = "3rdPlaceMedalImage"
+     let fourthPlaceMedalImage = "4thPlaceMedalImage"
+     let fifthPlaceMedalImage = "5thPlaceMedalImage"
+}
